Add test for new SelfPacedReadingInput and fix component
